fix: separate class to eth and nft
